refactor : station 삭제 관련 조건 추 (해당 역 참조 section이 있다면 삭제할 수 없도록)

+++ src/main/java/subway/service/StationService.java
@@ -1,8 +1,11 @@
 package subway.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import subway.dao.SectionDao;
 import subway.dao.StationDao;
 import subway.domain.station.Station;
+import subway.exception.InvalidStationException;
 import subway.exception.NotExistException;
 
 import java.util.List;
@@ -11,11 +14,14 @@ import java.util.stream.Collectors;
 @Service
 public class StationService {
     static final String NOT_EXIST_STATION_ERROR_MESSAGE = "해당 역이 존재하지 않습니다.";
+    static final String CANNOT_DELETE_STATION_ERROR_MESSAGE = "해당 역을 참조하는 구간이 존재합니다.";
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao,SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public Station createStation(String stationName) {
@@ -36,7 +42,11 @@ public class StationService {
         return station;
     }
 
+    @Transactional
     public void deleteStation(long id) {
+        if (!sectionDao.findByStationId(id).isEmpty()){
+            throw new InvalidStationException(CANNOT_DELETE_STATION_ERROR_MESSAGE);
+        }
         stationDao.deleteById(id);
     }
